+++ portal/tests.py
@@ -2,6 +2,8 @@ from django.contrib.auth.models import User
 from django.test import TestCase
 from django.urls import reverse
 
+from programs.models import Adult
+
 
 class PortalDashboardTests(TestCase):
     def setUp(self):
@@ -30,3 +32,28 @@ class PortalDashboardTests(TestCase):
         for s in problematic_strings:
             self.assertNotContains(response, f"{{#")
             self.assertNotContains(response, s)
+
+
+class PortalDashboardContactBoxRemovalTests(TestCase):
+    """The old 'Your Contact Information' box (name/phone/email) was removed."""
+
+    def setUp(self):
+        self.mentor = User.objects.create_user(
+            username="mentor_contact", password="password123"
+        )  # nosec B106
+        Adult.objects.create(
+            user=self.mentor,
+            first_name="Mentor",
+            last_name="User",
+            is_mentor=True,
+            phone_number="412-555-0100",
+        )
+        self.client.login(
+            username="mentor_contact", password="password123"
+        )  # nosec B106
+
+    def test_dashboard_does_not_show_contact_information_box(self):
+        response = self.client.get(reverse("profile_dashboard"))
+        self.assertEqual(response.status_code, 200)
+        self.assertNotContains(response, "Your Contact Information")
+        self.assertNotContains(response, "To update your contact information")